Revert "Feat/add status and type filters to filters list"
